requirements/requirements.txt
selenium
webdriver-manager
pydantic
python-dotenv